requirements/base.txt
django==3.1.1
psycopg2-binary==2.8.6
speedtest-cli==2.1.2